Update the rewards page
Seprate page for `Random Beacon` and `tBTC` instead of a single
`Overview` page.
